leave request

+++ database/migrations/2020_10_10_062827_create_leave_requests_table.php
@@ -20,7 +20,7 @@ class CreateLeaveRequestsTable extends Migration
             $table->UnsignedBigInteger('department_id');
             $table->dateTime('start_date');
             $table->dateTime('end_date');
-            $table->double('number_of_days');
+            $table->decimal('number_of_days', 10, 9);
             $table->UnsignedBigInteger('approver_role_id');
             $table->UnsignedBigInteger('status_id')->default(1);
             $table->timestamps();
